Move core and exceptions to process_implementations folder
