+++ openeo_processes_dask/process_implementations/core.py
@@ -6,7 +6,9 @@ from typing import Optional
 
 from openeo_pg_parser_networkx.pg_schema import ParameterReference
 
-from openeo_processes_dask.exceptions import ProcessParameterMissing
+from openeo_processes_dask.process_implementations.exceptions import (
+    ProcessParameterMissing,
+)
 
 logger = logging.getLogger(__name__)
