+++ openeo_processes_dask/process_implementations/cubes/_filter.py
@@ -2,8 +2,10 @@ from typing import Callable
 
 from openeo_pg_parser_networkx.pg_schema import BoundingBox, GeoJson, TemporalInterval
 
-from openeo_processes_dask.exceptions import DimensionNotAvailable
 from openeo_processes_dask.process_implementations.data_model import RasterCube
+from openeo_processes_dask.process_implementations.exceptions import (
+    DimensionNotAvailable,
+)
 
 
 def filter_spatial(data: RasterCube, geometries: GeoJson, **kwargs) -> RasterCube:
