+++ openeo_processes_dask/process_implementations/cubes/aggregate.py
@@ -3,10 +3,13 @@ from typing import Callable, Optional, Union
 import rasterio
 from openeo_pg_parser_networkx.pg_schema import TemporalInterval, TemporalIntervals
 
-from openeo_processes_dask.exceptions import DimensionNotAvailable, TooManyDimensions
 from openeo_processes_dask.process_implementations.data_model import (
     RasterCube,
     VectorCube,
+)
+from openeo_processes_dask.process_implementations.exceptions import (
+    DimensionNotAvailable,
+    TooManyDimensions,
 )
 
 __all__ = ["aggregate_temporal", "aggregate_temporal_period", "aggregate_spatial"]
