+++ openeo_processes_dask/process_implementations/cubes/merge.py
@@ -3,8 +3,10 @@ from typing import Callable, Optional
 import numpy as np
 import xarray as xr
 
-from openeo_processes_dask.exceptions import OverlapResolverMissing
 from openeo_processes_dask.process_implementations.data_model import RasterCube
+from openeo_processes_dask.process_implementations.exceptions import (
+    OverlapResolverMissing,
+)
 
 __all__ = ["merge_cubes"]
