+++ openeo_processes_dask/process_implementations/cubes/reduce.py
@@ -2,8 +2,10 @@ from typing import Callable, Optional
 
 import numpy as np
 
-from openeo_processes_dask.exceptions import DimensionNotAvailable
 from openeo_processes_dask.process_implementations.data_model import RasterCube
+from openeo_processes_dask.process_implementations.exceptions import (
+    DimensionNotAvailable,
+)
 
 __all__ = ["reduce_dimension", "reduce_spatial"]
