+++ tests/test_arrays.py
@@ -7,9 +7,12 @@ import pytest
 import xarray as xr
 from openeo_pg_parser_networkx.pg_schema import ParameterReference
 
-from openeo_processes_dask.exceptions import ArrayElementNotAvailable, TooManyDimensions
 from openeo_processes_dask.process_implementations.arrays import *
 from openeo_processes_dask.process_implementations.cubes.reduce import reduce_dimension
+from openeo_processes_dask.process_implementations.exceptions import (
+    ArrayElementNotAvailable,
+    TooManyDimensions,
+)
 from tests.general_checks import general_output_checks
 from tests.mockdata import create_fake_rastercube
