+++ tests/test_core.py
@@ -1,8 +1,10 @@
 import pytest
 from openeo_pg_parser_networkx.pg_schema import ParameterReference
 
-from openeo_processes_dask.core import process
-from openeo_processes_dask.exceptions import ProcessParameterMissing
+from openeo_processes_dask.process_implementations.core import process
+from openeo_processes_dask.process_implementations.exceptions import (
+    ProcessParameterMissing,
+)
 
 
 def test_process_decorator():
