+++ tests/test_merge.py
@@ -5,11 +5,13 @@ import pytest
 import xarray as xr
 from openeo_pg_parser_networkx.pg_schema import ParameterReference
 
-from openeo_processes_dask.exceptions import OverlapResolverMissing
 from openeo_processes_dask.process_implementations import merge_cubes
 from openeo_processes_dask.process_implementations.cubes.merge import (
     NEW_DIM_COORDS,
     NEW_DIM_NAME,
+)
+from openeo_processes_dask.process_implementations.exceptions import (
+    OverlapResolverMissing,
 )
 from tests.mockdata import create_fake_rastercube
